Update README.md
GlobalVison is an app that processes an image and recognizes and predicts whats in the picture, and then it says it out loud. GlobalVison uses an API database to process any image and identify it, along with utilizing text-to-speech to say the name of the image out loud. This product is built not only for anyone willing to try this new technology but also especially for those with impaired vision, giving them a chance to see the world. Although GlobalVision is not something that is necessarily ready to be incorporated into the daily lives of blind people, it is a step in the right direction to advocate for those with disabilities and push for positive change utilizing AI technology, giving everyone a chance to see.